Struct smartstring::Compact

pub struct Compact;

A compact string representation equal to [String] in size with guaranteed inlining.

This representation relies on pointer alignment to be able to store a discriminant bit in its inline form that will never be present in its [String] form, thus giving us 24 bytes on 64-bit architectures, and 12 bytes on 32-bit, minus one bit, to encode our inline string. It uses the rest of the discriminant bit’s byte to encode the string length, and the remaining bytes (23 or 11 depending on arch) to store the string data. When the available space is exceeded, it swaps itself out with a [String] containing its previous contents, relying on the discriminant bit in the [String]’s pointer to be unset, so we can store the [String] safely without taking up any extra space for a discriminant.

This performs generally as well as [String] on all ops on boxed strings, and better than [String]s on inlined strings.