Struct smartstring::SmartString

#[repr(C)]

#[repr(align(8))]
pub struct SmartString<Mode: SmartStringMode> { /* fields omitted */ }

A smart string.

This wraps one of two string types: an inline string or a boxed string. Conversion between the two happens opportunistically and transparently.

It takes a layout as its type argument: one of Compact or LazyCompact.

It mimics the interface of String except where behaviour cannot be guaranteed to stay consistent between its boxed and inline states. This means you still have capacity() and shrink_to_fit(), relating to state that only really exists in the boxed variant, because the inline variant can still give sensible behaviour for these operations, but with_capacity(), reserve() etc are absent, because they would have no effect on inline strings and the requested state changes wouldn’t carry over if the inline string is promoted to a boxed one - not without also storing that state in the inline representation, which would waste precious bytes for inline string data.

Implementations

impl SmartString<LazyCompact>

pub const fn new_const() -> Self

Construct an empty string.

This is a const fn version of SmartString::new. It’s a temporary measure while we wait for trait bounds on type arguments to const fns to stabilise, and will be deprecated once this happens.

impl SmartString<Compact>

pub const fn new_const() -> Self

Construct an empty string.

This is a const fn version of SmartString::new. It’s a temporary measure while we wait for trait bounds on type arguments to const fns to stabilise, and will be deprecated once this happens.

impl<Mode: SmartStringMode> SmartString<Mode>

pub fn new() -> Self

Construct an empty string.

pub fn len(&self) -> usize

Return the length in bytes of the string.

Note that this may differ from the length in chars.

pub fn is_empty(&self) -> bool

Test whether the string is empty.

pub fn is_inline(&self) -> bool

Test whether the string is currently inlined.

pub fn as_str(&self) -> &str

Get a reference to the string as a string slice.

pub fn as_mut_str(&mut self) -> &mut str

Get a reference to the string as a mutable string slice.

pub fn push(&mut self, ch: char)

Push a character to the end of the string.

pub fn push_str(&mut self, string: &str)

Copy a string slice onto the end of the string.

pub fn capacity(&self) -> usize

Return the currently allocated capacity of the string.

Note that if this is a boxed string, it returns String::capacity(), but an inline string always returns MAX_INLINE.

Note also that if a boxed string is converted into an inline string, its capacity is deallocated, and if the inline string is promoted to a boxed string in the future, it will be reallocated with a default capacity.

pub fn shrink_to_fit(&mut self)

Shrink the capacity of the string to fit its contents exactly.

This has no effect on inline strings, which always have a fixed capacity. Thus, it’s not safe to assume that capacity() will equal len() after calling this.

Calling this on a LazyCompact string that is currently heap allocated but is short enough to be inlined will deallocate the heap allocation and convert it to an inline string.

pub fn truncate(&mut self, new_len: usize)

Truncate the string to new_len bytes.

If new_len is larger than the string’s current length, this does nothing. If new_len isn’t on a UTF-8 character boundary, this method panics.

pub fn pop(&mut self) -> Option<char>

Pop a char off the end of the string.

pub fn remove(&mut self, index: usize) -> char

Remove a char from the string at the given index.

If the index doesn’t fall on a UTF-8 character boundary, this method panics.

pub fn insert(&mut self, index: usize, ch: char)

Insert a char into the string at the given index.

If the index doesn’t fall on a UTF-8 character boundary, this method panics.

pub fn insert_str(&mut self, index: usize, string: &str)

Insert a string slice into the string at the given index.

If the index doesn’t fall on a UTF-8 character boundary, this method panics.

pub fn split_off(&mut self, index: usize) -> Self

Split the string into two at the given index.

Returns the content to the right of the index as a new string, and removes it from the original.

If the index doesn’t fall on a UTF-8 character boundary, this method panics.

pub fn clear(&mut self)

Clear the string.

This causes any memory reserved by the string to be immediately deallocated.

pub fn retain<F>(&mut self, f: F) where
F: FnMut(char) -> bool,

Filter out chars not matching a predicate.

pub fn drain<R>(&mut self, range: R) -> Drain<'_, Mode> where
R: RangeBounds<usize>,

Construct a draining iterator over a given range.

This removes the given range from the string, and returns an iterator over the removed chars.

pub fn replace_range<R>(&mut self, range: R, replace_with: &str) where
R: RangeBounds<usize>,

Replaces a range with the contents of a string slice.