Trait smartstring::SmartStringMode

pub trait SmartStringMode {
    type BoxedString: BoxedString + From<String>;
    type InlineArray: AsRef<[u8]> + AsMut<[u8]> + Clone + Copy;

    const DEALLOC: bool;
}

Marker trait for SmartString representations.

See LazyCompact and Compact.

Associated Types

type BoxedString: BoxedString + From<String>

The boxed string type for this layout.

type InlineArray: AsRef<[u8]> + AsMut<[u8]> + Clone + Copy

The inline string type for this layout.

Associated Constants

const DEALLOC: bool

A constant to decide whether to turn a wrapped string back into an inlined string whenever possible (true) or leave it as a wrapped string once wrapping has occurred (false).